# src/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Data API, database migration scripts, and backend services for all Podverse models.",
    "version": "1.0.0",
    "title": "Podverse API",
    "termsOfService": "https://podverse.fm/terms/",
    "contact": {
      "email": "dev@podverse.fm"
    },
    "license": {
      "name": "GNU Affero General Public License v3",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
  "host": "localhost:1234",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/podcast": {
      "get": {
        "tags": [
          "podcast"
        ],
        "summary": "Find podcasts by fields",
        "description": "Returns an array of podcasts",
        "operationId": "getPodcasts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authors",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "categories",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "type": "string"
          },
          {
            "name": "episodes",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "feedLastUpdated",
            "in": "query",
            "type": "string",
            "format": "date"
          },
          {
            "name": "guid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "imageUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isExplicit",
            "in": "query",
            "type": "boolean"
          },
          {
            "name": "language",
            "in": "query",
            "type": "string"
          },
          {
            "name": "linkUrl",
            "in": "query",
            "type": "string"
          },
          {
            "name": "title",
            "in": "query",
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Podcast"
            }
          }
        }
      }
    },
    "/podcast/{podcastId}": {
      "get": {
        "tags": [
          "podcast"
        ],
        "summary": "Find podcast by ID",
        "description": "Returns a single podcast",
        "operationId": "getPodcastById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "podcastId",
            "in": "path",
            "description": "ID of podcast to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Podcast"
            }
          },
          "404": {
            "description": "Podcast not found"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Find users by fields",
        "description": "Returns an array of users",
        "operationId": "getUsers",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Add a user",
        "description": "Returns the new user",
        "operationId": "createUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "example": {
                "email": "string - required",
                "name": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/user/{userId}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Find user by ID",
        "description": "Returns a single user",
        "operationId": "getUserById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete user by ID",
        "description": "Returns success or error message",
        "operationId": "getUserById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "authorization": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Author": {
      "type": "object",
      "required": [
        "id",
        "name",
        "slug"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        }
      }
    },
    "Category": {
      "type": "object",
      "required": [
        "id",
        "slug",
        "title"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        }
      }
    },
    "Episode": {
      "type": "object",
      "required": [
        "id",
        "mediaUrl"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "mediaUrl": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "duration": {
          "type": "integer"
        },
        "episodeType": {
          "type": "string"
        },
        "guid": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "isExplicit": {
          "type": "boolean"
        },
        "isPublic": {
          "type": "boolean"
        },
        "linkUrl": {
          "type": "string"
        },
        "mediaFilesize": {
          "type": "integer"
        },
        "mediaType": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "pubDate": {
          "type": "string",
          "format": "date"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "mediaRefs": {
          "$ref": "#/definitions/MediaRef"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "FeedUrl": {
      "type": "object",
      "required": [
        "url",
        "podcast"
      ],
      "properties": {
        "url": {
          "type": "string"
        },
        "isAuthority": {
          "type": "boolean"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "MediaRef": {
      "type": "object",
      "required": [
        "id",
        "episodeMediaUrl",
        "startTime"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "episodeMediaUrl": {
          "type": "string"
        },
        "startTime": {
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "endTime": {
          "type": "integer"
        },
        "episodeDuration": {
          "type": "integer"
        },
        "episodeGuid": {
          "type": "string"
        },
        "episodeId": {
          "type": "string"
        },
        "episodeImageUrl": {
          "type": "string"
        },
        "episodeLinkUrl": {
          "type": "string"
        },
        "episodePubDate": {
          "type": "string",
          "format": "date"
        },
        "episodeSummary": {
          "type": "string"
        },
        "episodeTitle": {
          "type": "string"
        },
        "isPublic": {
          "type": "boolean"
        },
        "ownerId": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "podcastFeedUrl": {
          "type": "string"
        },
        "podcastGuid": {
          "type": "string"
        },
        "podcastId": {
          "type": "string"
        },
        "podcastImageUrl": {
          "type": "string"
        },
        "podcastIsExplicit": {
          "type": "boolean"
        },
        "podcastTitle": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "episode": {
          "$ref": "#/definitions/Episode"
        },
        "owner": {
          "$ref": "#/definitions/User"
        },
        "podcast": {
          "$ref": "#/definitions/Podcast"
        }
      }
    },
    "Playlist": {
      "type": "object",
      "required": [
        "id",
        "owner"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "isMyClips": {
          "type": "boolean"
        },
        "isPublic": {
          "type": "boolean"
        },
        "itemsOrder": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "mediaRefs": {
          "$ref": "#/definitions/MediaRef"
        },
        "owner": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "Podcast": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "feedLastUpdated": {
          "type": "string",
          "format": "date"
        },
        "guid": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "isExplicit": {
          "type": "boolean"
        },
        "language": {
          "type": "string"
        },
        "linkUrl": {
          "type": "string"
        },
        "pastAllTimeTotalUniquePageviews": {
          "type": "integer"
        },
        "pastHourTotalUniquePageviews": {
          "type": "integer"
        },
        "pastDayTotalUniquePageviews": {
          "type": "integer"
        },
        "pastWeekTotalUniquePageviews": {
          "type": "integer"
        },
        "pastMonthTotalUniquePageviews": {
          "type": "integer"
        },
        "pastYearTotalUniquePageviews": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "authors": {
          "$ref": "#/definitions/Author"
        },
        "categories": {
          "$ref": "#/definitions/Category"
        },
        "episodes": {
          "$ref": "#/definitions/Episode"
        },
        "feedUrls": {
          "$ref": "#/definitions/FeedUrl"
        }
      }
    },
    "User": {
      "type": "object",
      "required": [
        "id",
        "email"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "subscribedPodcastIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "playlists": {
          "$ref": "#/definitions/Playlist"
        }
      }
    }
  }
}